+++ src/controllers/page.js
@@ -1,8 +1,6 @@
 const SHOWING_FILMS_COUNT_ON_START = 5;
 const SHOWING_FILMS_COUNT_BY_BUTTON = 5;
 
-import FilmComponent from "../components/film";
-import FilmPopupComponent from "../components/film-details";
 import ShowMoreComponent from "../components/show-more";
 import FilmExtraListComponent from "../components/film-extra-list";
 import NoFilmsComponent from "../components/no-films";
@@ -10,46 +8,15 @@ import FilterComponent from "../components/filter";
 import FilmListComponent from "../components/film-list";
 import {render, Position} from "../utils";
 import {films, filmsRated, filmsCommented} from "../data";
+import MovieController from "./movie";
 
-const siteMainElement = document.querySelector(`.main`);
+const renderFilms = (movies, listFilms) => {
+  return movies.map((movie) => {
+    const movieController = new MovieController(listFilms);
+    movieController.render(movie);
 
-const renderFilm = (film, container) => {
-  const filmComponent = new FilmComponent(film);
-  const filmPopupComponent = new FilmPopupComponent(film);
-
-  const showPopup = () => {
-    render(siteMainElement, filmPopupComponent.getElement(), Position.BEFOREEND);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      siteMainElement.removeChild(filmPopupComponent.getElement());
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  filmComponent.setTitleClickHandler(() => {
-    showPopup();
-    document.addEventListener(`keydown`, onEscKeyDown);
+    return movieController;
   });
-
-  filmComponent.setPosterClickHandler(() => {
-    showPopup();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  filmComponent.setCommentsClickHandler(() => {
-    showPopup();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  filmPopupComponent.setCloseClickHandler(() => {
-    siteMainElement.removeChild(filmPopupComponent.getElement());
-  });
-
-  render(container, filmComponent.getElement(), Position.BEFOREEND);
 };
 
 export default class PageController {
@@ -75,17 +42,20 @@ export default class PageController {
 
       const filmsList = document.querySelector(`.films-list`);
       const filmListContainer = document.querySelector(`.films-list__container`);
+
       const filterButtons = this._filterComponent.getElement().querySelectorAll(`.sort__button`);
+
       const activeFilter = (filter) => {
         Array.from(filterButtons).filter((button) => {
           return button.dataset.sortType === filter;
         })[0].classList.add(`sort__button--active`);
-      }
+      };
 
       this._filterComponent.setSortTypeChangeHandler((sortType) => {
         filterButtons.forEach((button) => {
           button.classList.remove(`sort__button--active`);
         });
+
         switch (sortType) {
           case `date`:
             filmsData = this._films.slice().sort((a, b) => new Date(a.year) - new Date(b.year));
@@ -102,14 +72,10 @@ export default class PageController {
         }
 
         filmListContainer.innerHTML = null;
-        filmsData.slice(0, showingFilmsCount).forEach((film) => {
-          renderFilm(film, filmListContainer);
-        });
+        renderFilms(filmsData.slice(0, showingFilmsCount), filmListContainer);
       });
 
-      filmsData.slice(0, showingFilmsCount).forEach((film) => {
-        renderFilm(film, filmListContainer);
-      });
+      renderFilms(filmsData.slice(0, showingFilmsCount), filmListContainer);
 
       if (filmsData.length > SHOWING_FILMS_COUNT_ON_START) {
         render(filmsList, this._showMoreComponent.getElement(), Position.BEFOREEND);
@@ -124,13 +90,8 @@ export default class PageController {
           `.films-list--extra .films-list__container`
       );
 
-      filmsRated.forEach((film) => {
-        renderFilm(film, filmsExtraElements[0]);
-      });
-
-      filmsCommented.forEach((film) => {
-        renderFilm(film, filmsExtraElements[1]);
-      });
+      renderFilms(filmsRated, filmsExtraElements[0]);
+      renderFilms(filmsCommented, filmsExtraElements[1]);
     }
 
     this._showMoreComponent.setClickHandler(() => {
@@ -139,10 +100,7 @@ export default class PageController {
       showingFilmsCount = showingFilmsCount + SHOWING_FILMS_COUNT_BY_BUTTON;
       const filmListContainer = document.querySelector(`.films-list__container`);
 
-      filmsData.slice(prevTasksCount, showingFilmsCount)
-        .forEach((film) => {
-          renderFilm(film, filmListContainer);
-        });
+      renderFilms(filmsData.slice(prevTasksCount, showingFilmsCount), filmListContainer);
 
       if (showingFilmsCount >= filmsData.length) {
         this._showMoreComponent.getElement().remove();
